προσθήκη alert μηνυμάτων

+++ frontend/src/app/components/admin/users/users.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { DataService } from '../../../services/data/data.service';
+import { HelperService } from '../../../services/helper/helper.service';
 
 @Component({
   selector: 'app-users',
@@ -9,7 +10,7 @@ import { DataService } from '../../../services/data/data.service';
 export class UsersComponent {
 
   users: any = [];
-  constructor(public data: DataService) {
+  constructor(public data: DataService, private helper: HelperService) {
     this.getAllUsers();
     if(this.data.permissions.length === 0) {
       this.data.getAllPermissions();
@@ -20,7 +21,6 @@ export class UsersComponent {
     this.data.getAllUsers().then( response => {
       this.users = response;
     }).catch( error => {
-      alert('something error')
     })
   }
 
@@ -33,10 +33,15 @@ export class UsersComponent {
   }
 
   deleteUser(user: any) {
-    this.data.deleteUser(user).then( (response: any) => {
-      if(response === 'success') {
-        this.getAllUsers();
+    this.helper.presentAlert('question', 'alert.deleteQuestion', 'alert.deleteUserQuestion', true).then((alert: any) => {
+      console.log(alert)
+      if(alert.isConfirmed) {
+        this.data.deleteUser(user).then( (response: any) => {
+          if(response === 'success') {
+            this.getAllUsers();
+          }
+        });
       }
-    })
+    });
   }
 }
